Create Email-Text-to-Google-Sheets
This VBScript will extract data in the format "Title1/Title2/Title3/Title4/Title5." "Titles" can be renamed to describe whatever Column heading you are going for. You'll need to have the emails you are trying to have picked up, given a specific Subject Line, and set a rule in your email to catch that phrase.